g: add rx_handled into USB_STREAM_CONFIG
This patch add rx_handle into USB_STREAM_CONFIG, which tracks
how many of the bytes in the HW FIFO was moved into the incoming
queue. This used to be defined as local static variable for multiple
USB_STREAM_CONFIGs.

BUG=None
BRANCH=None
TEST=manually ran uart_stress_tester.sh and flash_ec on Bob.

+++ chip/g/usb_console.c
@@ -120,7 +120,7 @@ static void rx_fifo_handler(void)
 	 * and ep_buf_rx remains valid and unchanged until software tells the
 	 * the hardware engine to accept more input.
 	 */
-	int rx_in_fifo, rx_left;
+	size_t rx_in_fifo, rx_left;
 
 	/*
 	 * The rx_handled variable tracks how many of the bytes in the HW FIFO
@@ -129,7 +129,7 @@ static void rx_fifo_handler(void)
 	 * the next time this function is called we can try to shove the rest
 	 * of the HW FIFO bytes into the queue.
 	 */
-	static int rx_handled;
+	static size_t rx_handled;
 
 	/* If the HW FIFO isn't ready, then we're waiting for more bytes */
 	if (!rx_fifo_is_ready())
